nSize shall presumably be of type long, not double
It was introduced with 6f660293885733e64602aa6144c36ed7e7b43faf "implement
VclVPaned and decripple formula panel", but using double was presumably a typo,
as it is initialized with a long value (Splitter::GetSplitPosPixel()), is
combined with other long values (Size::Height()), and passed to long
nSecondHeight parameter of VclVPaned::arrange.

+++ vcl/source/window/layout.cxx
@@ -2501,7 +2501,7 @@ void VclVPaned::dispose()
 
 IMPL_LINK(VclVPaned, SplitHdl, Splitter*, pSplitter, void)
 {
-    double nSize = pSplitter->GetSplitPosPixel();
+    long nSize = pSplitter->GetSplitPosPixel();
     Size aSplitterSize(m_pSplitter->GetSizePixel());
     Size aAllocation(GetSizePixel());
     arrange(aAllocation, nSize, aAllocation.Height() - nSize - aSplitterSize.Height());
